+++ docs/conf.py
@@ -19,10 +19,14 @@ sys.path.insert(0, os.path.abspath('..'))
 #
 # needs_sphinx = '1.0'
 
+package_root = "."
+
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc',
+extensions = ['sphinxcontrib.extras_require',
+              'sphinx-prompt',
+              'sphinx.ext.autodoc',
               'sphinx.ext.intersphinx',
               'sphinx.ext.todo',
               'sphinx.ext.coverage',