+++ CHANGELOG.md
@@ -17,7 +17,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Maintenance
   * Reduce duplication of code in instrument modules
   * Include flake8 linting of docstrings and style in Github Actions
-  * Move OMNI HRO custom functions to an instrument module
+  * Move OMNI HRO custom functions to a methods module
+  * Deprecate OMNI HRO custom functions in instrument module
 
 ## [0.0.3] - 2022-05-18
 * Include flake8 linting of docstrings and style in Github Actions